+++ stdnet/dispatch/__init__.py
@@ -1,0 +1,6 @@
+"""Multi-consumer multi-producer dispatching mechanism
+
+Originally based on pydispatch (BSD) http://pypi.python.org/pypi/PyDispatcher/2.0.1
+See license.txt for original license.
+"""
+from stdnet.dispatch.dispatcher import Signal